+++ src/lightning/pytorch/plugins/precision/fsdp.py
@@ -112,7 +112,10 @@ class FSDPPrecisionPlugin(PrecisionPlugin):
             buffer_dtype=buffer_dtype,
         )
 
-    def init_context(self) -> ContextManager:
+    def tensor_init_context(self) -> ContextManager:
+        return _DtypeContextManager(self._desired_input_dtype)
+
+    def module_init_context(self) -> ContextManager:
         return _DtypeContextManager(self.mixed_precision_config.param_dtype or torch.float32)
 
     def forward_context(self) -> ContextManager: